requirements.txt
discord.py==2.3.2
aiohttp==3.9.1
python-dotenv==1.0.0
aiosqlite==0.19.0
cc-ampapi==1.3.0
loguru==0.7.2
